feat: update pool card
