+++ devlink/FundingPoolTeaser.module.css
@@ -287,8 +287,8 @@
 
 .sponsor-image._50.floating {
   position: absolute;
-  left: 2%;
-  top: 2%;
+  left: 4%;
+  top: 4%;
   right: auto;
   bottom: auto;
