+++ devlink/global.css
@@ -17054,8 +17054,7 @@ blockquote {
 
 .what-if-title {
   width: 80%;
-  margin-bottom: 20px;
-  font-size: 26px;
+  font-size: 20px;
   line-height: 30px;
   text-align: center;
 }
